+++ tests/Unit/Services/TestOutputRenderer/StepRendererTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace webignition\BasilRunner\Tests\Unit\Services\ResultPrinter\StepRenderer;
+namespace webignition\BasilRunner\Tests\Unit\Services\TestOutputRenderer;
 
 use PHPUnit\Runner\BaseTestRunner;
 use webignition\BaseBasilTestCase\BasilTestCaseInterface;
@@ -14,7 +14,7 @@ use webignition\BasilRunner\Services\ResultPrinter\ConsoleOutputFactory;
 use webignition\BasilRunner\Services\ResultPrinter\FailedAssertion\SummaryHandler;
 use webignition\BasilRunner\Services\ResultPrinter\FailedAssertion\SummaryLineFactory;
 use webignition\BasilRunner\Services\ResultPrinter\StatementLineRenderer;
-use webignition\BasilRunner\Services\ResultPrinter\StepRenderer\StepRenderer;
+use webignition\BasilRunner\Services\TestOutputRenderer\StepRenderer;
 use webignition\BasilRunner\Tests\Unit\AbstractBaseTest;
 
 class StepRendererTest extends AbstractBaseTest
